requirements.txt
django==3.2.9
django-cors-headers
django-environ==0.8.1
django-storages==1.12.3
djangorestframework==3.12.4
social-auth-app-django
